requirements.txt
fastapi
uvicorn
nemoguardrails
google-generativeai
requests
